feat: add newsletter subscription feature with public POST and superadmin GET API
